+++ Reddit/Easy/making_links/src/main.rs
@@ -2,5 +2,4 @@ extern crate making_links;
 
 fn main() {
     println!("{}", making_links::convert("7026425611433322325"));
-
 }